[496978] Adapt HTML content to markup when using paste command
Preprocess any HTML content found in the clipboard by converting it to
the current marking. To do so we have to strip some styling information
and ignore a few synthetic spans.

This change add a dependency from wikitext.ui to the core plugin of the
HTML language markup to enable the editor to parse the HTML language.
This added dependency has no visible impact in an environment as this
plugisn does not make any contribution to the UI.

Bug: 496978
Change-Id: Id9cd0fdbcdc5823ca09226507de7b44cdaa80dcf
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>